# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para o projeto Leinadium/puc-disciplinas",
        "title": "puc-disciplinas API",
        "contact": {
            "name": "Daniel",
            "url": "https://github.com/Leinadium/puc-disciplinas"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/disciplina/info": {
            "get": {
                "description": "Coleta a informação completa de uma disciplina",
                "produces": [
                    "application/json"
                ],
                "summary": "Coleta a informação completa de uma disciplina",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Codigo da disciplina",
                        "name": "c",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/disciplinas.ResultInfo"
                            }
                        }
                    },
                    "404": {
                        "description": "Disciplina nao encontrada",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao executar query",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/pesquisa/cursadas": {
            "get": {
                "description": "Coleta os codigos das disciplinas que o usuario cursou",
                "produces": [
                    "application/json"
                ],
                "summary": "Coleta os codigos das disciplinas que o usuario cursou",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/disciplinas.ResultCodigo"
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario nao logado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao executar query",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pesquisa/faltacursar": {
            "get": {
                "description": "Coleta o código das disciplinas que o usuário falta cursar. É vazio se o usuário nao estiver logado.",
                "produces": [
                    "application/json"
                ],
                "summary": "Coleta as disciplinas que o usuario falta cursar",
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/disciplinas.ResultCodigo"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao executar query",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pesquisa/info": {
            "get": {
                "description": "Coleta as informações das disciplinas, além da data de modificação.",
                "produces": [
                    "application/json"
                ],
                "summary": "Coleta todas as disciplinas",
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/disciplinas.ResultLista"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao executar query",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pesquisa/podecursar": {
            "get": {
                "description": "Coleta o código das disciplinas que o usuário pode cursar",
                "produces": [
                    "application/json"
                ],
                "summary": "Coleta as disciplinas que o usuario pode cursar",
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/disciplinas.ResultCodigo"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao executar query",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "disciplinas.ResultCodigo": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                }
            }
        },
        "disciplinas.ResultInfo": {
            "type": "object",
            "properties": {
                "codDiscDepen": {
                    "type": "string"
                },
                "codigo": {
                    "type": "string"
                },
                "creditos": {
                    "type": "integer"
                },
                "ementa": {
                    "type": "string"
                },
                "grupoPrereq": {
                    "type": "integer"
                },
                "mediaAvaliacao": {
                    "type": "number"
                },
                "mediaGrau": {
                    "type": "number"
                },
                "nome": {
                    "type": "string"
                },
                "qtdAlunos": {
                    "type": "integer"
                },
                "qtdAvaliacoes": {
                    "type": "integer"
                }
            }
        },
        "disciplinas.ResultLista": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "creditos": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "qtdTurmas": {
                    "type": "integer"
                },
                "qtdVagas": {
                    "type": "integer"
                }
            }
        },
...
    }
}